Uservar access
Now all of the user vars you can define in the user file (for example Real Name, Mail address, etc.) can be actually used :).

In the blueprint you can now define fields to automatically fill with user vars, for example:

title: Page
pages: true
files: true
fields:
  title:
    label: Title
    type:  text
  text:
    label: Text
    type:  textarea
    size:  large
  author:
    uservar: Name

After defining that (you don't need to specify a type), a new hidden text field with the user var as value will be added to the form and when you click "Save", this var will be saved to the content file.

In my case this would add a new "Author" field in the content file with the value "Lukas Bestle".

This is very cool to automatically set the author (-> last editing of an article) to a person without having to manually set this every time.