Made "uservars" a lot better
The idea Bastian had to use something like
{{user.name}}
is now possible for any field.

You can just write something like above into the "default" attribute of any field (hidden or not) to access the user var with the given name and show it by default.

That's actually very much better :).