Update lib/data.php
Creating a new file in a multi-lingual setup with the panel used to create an empty file for every available languages, which caused non-yet-translated page's content to be empty when users would visit the website. It now creates a file only for the default language. This is a saner and simpler default.